define event interface use Event Queue
