+++ Libraries/Opc.Ua.Server/Subscription/MonitoredItem/IMonitoredItem.cs
