+++ Libraries/Opc.Ua.Server/Subscription/MonitoredItem/IMonitoredItemQueue.cs
@@ -32,19 +32,19 @@ using System;
 namespace Opc.Ua.Server
 {
     /// <summary>
-    /// Used to create <see cref="IMonitoredItemQueue{T}"/> and dispose unmanaged resources on server shutdown
+    /// Used to create <see cref="IDataChangeMonitoredItemQueue"/> / <see cref="IEventMonitoredItemQueue"/> and dispose unmanaged resources on server shutdown
     /// </summary>
     public interface IMonitoredItemQueueFactory : IDisposable
     {
         /// <summary>
         /// Creates an empty queue for data values.
         /// </summary>
-        IMonitoredItemQueue<DataValue> CreateDataValueQueue(bool isDurable);
+        IDataChangeMonitoredItemQueue CreateDataValueQueue(bool isDurable);
 
         /// <summary>
         /// Creates an empty queue for events.
         /// </summary>
-        IMonitoredItemQueue<EventFieldList> CreateEventQeue(uint monitoredItemId, bool isDurable);
+        IEventMonitoredItemQueue CreateEventQueue(bool isDurable);
 
         /// <summary>
         /// If true durable queues can be created by the factory, if false only regular queues with small queue sizes are returned
@@ -56,7 +56,7 @@ namespace Opc.Ua.Server
     /// Provides a durable queue for data changes and events that can handle a queue size of several thousand elements.
     /// T defines if the queue is for events or for DataValues
     /// </summary>
-    public interface IMonitoredItemQueue<T> : IDisposable
+    public interface IDataChangeMonitoredItemQueue : IDisposable
     {
         /// <summary>
         /// True if the queue is in durable mode and persists the queue values / supports a large queue size
@@ -85,20 +85,7 @@ namespace Opc.Ua.Server
         /// </summary>
         /// <param name="value">The value to queue.</param>
         /// <param name="error">The error to queue.</param>
-        void QueueValue(T value, ServiceResult error);
-
-        /// <summary>
-        /// Replace the newest value in the queue with the provided Value. Used when values are provided faster than the sampling interval
-        /// </summary>
-        /// <param name="value">The value to queue.</param>
-        /// <param name="error">The error to queue.</param>
-        void OverwriteLastValue(T value, ServiceResult error);
-
-        /// <summary>
-        /// Returns the last value in the queue without dequeuing
-        /// </summary>
-        /// <returns>the last value, null if queue is empty</returns>
-        T PeekLastValue();
+        void Enqueue(DataValue value, ServiceResult error);
 
         /// <summary>
         /// Dequeue the oldest value in the queue.
@@ -106,6 +93,69 @@ namespace Opc.Ua.Server
         /// <param name="value">The value.</param>
         /// <param name="error">The error associated with the value.</param>
         /// <returns>True if a value was found. False if the queue is empty.</returns>
-        bool Dequeue(out T value, out ServiceResult error);
+        bool Dequeue(out DataValue value, out ServiceResult error);
+
+        /// <summary>
+        /// Replace the newest value in the queue with the provided Value. Used when values are provided faster than the sampling interval
+        /// </summary>
+        /// <param name="value">The value to queue.</param>
+        /// <param name="error">The error to queue.</param>
+        void OverwriteLastValue(DataValue value, ServiceResult error);
+
+        /// <summary>
+        /// Returns the last value in the queue without dequeuing
+        /// </summary>
+        /// <returns>the last value, null if queue is empty</returns>
+        DataValue PeekLastValue();
+    }
+
+    /// <summary>
+    /// Provides a durable queue for data changes and events that can handle a queue size of several thousand elements.
+    /// T defines if the queue is for events or for DataValues
+    /// </summary>
+    public interface IEventMonitoredItemQueue : IDisposable
+    {
+        /// <summary>
+        /// True if the queue is in durable mode and persists the queue values / supports a large queue size
+        /// </summary>
+        bool IsDurable { get; }
+
+        /// <summary>
+        /// Gets the current queue size.
+        /// </summary>
+        uint QueueSize { get; }
+
+        /// <summary>
+        /// Gets number of elements actually contained in value queue.
+        /// </summary>
+        int ItemsInQueue { get; }
+
+        /// <summary>
+        /// Sets the queue size.
+        /// </summary>
+        /// <param name="queueSize">The new queue size.</param>
+        /// <param name="discardOldest">if true remove oldest entries from the queue when the queue size decreases, else remove newest</param>
+        void SetQueueSize(uint queueSize, bool discardOldest);
+
+        /// <summary>
+        /// Checks the last 1k queue entries if the event is already in there
+        /// </summary>
+        /// <param name="instance">the event to chack for duplicates</param>
+        /// <returns>true if event already in queue</returns>
+        bool IsEventContainedInQueue(IFilterTarget instance);
+
+
+        /// <summary>
+        /// Adds the value to the queue.
+        /// </summary>
+        /// <param name="value">The value to queue.</param>
+        void Enqueue(EventFieldList value);
+
+        /// <summary>
+        /// Dequeue the oldest value in the queue.
+        /// </summary>
+        /// <param name="value">The value.</param>
+        /// <returns>True if a value was found. False if the queue is empty.</returns>
+        bool Dequeue(out EventFieldList value);
     }
 }
