+++ Libraries/Opc.Ua.Server/Subscription/MonitoredItem/MonitoredItem.cs
@@ -470,9 +470,9 @@ namespace Opc.Ua.Server
                         return m_eventQueueHandler.ItemsInQueue;
                     }
 
-                    if (m_queueHandler != null)
+                    if (m_dataValueQueueHandler != null)
                     {
-                        return m_queueHandler.ItemsInQueue;
+                        return m_dataValueQueueHandler.ItemsInQueue;
                     }
 
                     return 0;
@@ -876,7 +876,7 @@ namespace Opc.Ua.Server
         {
             if (m_queueSize > 1)
             {
-                m_queueHandler.QueueValue(value, error);
+                m_dataValueQueueHandler.QueueValue(value, error);
             }
 
             if (m_lastValue != null)
@@ -1207,17 +1207,17 @@ namespace Opc.Ua.Server
                 m_readyToPublish = false;
 
                 // check if queueing enabled.
-                if (m_queueHandler != null && (!m_resendData || m_queueHandler.ItemsInQueue != 0))
+                if (m_dataValueQueueHandler != null && (!m_resendData || m_dataValueQueueHandler.ItemsInQueue != 0))
                 {
                     DataValue value = null;
                     ServiceResult error = null;
 
-                    while (m_queueHandler.Publish(out value, out error))
+                    while (m_dataValueQueueHandler.PublishSingleValue(out value, out error))
                     {
                         Publish(context, notifications, diagnostics, value, error);
                         if (m_resendData)
                         {
-                            m_readyToPublish = m_queueHandler.ItemsInQueue > 0;
+                            m_readyToPublish = m_dataValueQueueHandler.ItemsInQueue > 0;
                             break;
                         }
                     }
@@ -1661,8 +1661,8 @@ namespace Opc.Ua.Server
                 {
                     m_eventQueueHandler?.Dispose();
                     m_eventQueueHandler = null;
-                    m_queueHandler?.Dispose();
-                    m_queueHandler = null;
+                    m_dataValueQueueHandler?.Dispose();
+                    m_dataValueQueueHandler = null;
                     break;
                 }
 
@@ -1688,34 +1688,34 @@ namespace Opc.Ua.Server
                     {
                         if (m_queueSize <= 1)
                         {
-                            m_queueHandler?.Dispose();
-                            m_queueHandler = null;
-                            m_eventQueueHandler?.Dispose();
+                            Utils.SilentDispose(m_dataValueQueueHandler);
+                            m_dataValueQueueHandler = null;
+                            Utils.SilentDispose(m_eventQueueHandler);
                             m_eventQueueHandler = null;
                             break; // queueing is disabled
                         }
 
                         bool queueLastValue = false;
 
-                        if (m_queueHandler == null)
+                        if (m_dataValueQueueHandler == null)
                         {
-                            m_queueHandler = new QueueHandler(Id, IsDurable, m_monitoredItemQueueFactory, QueueOverflowHandler);
+                            m_dataValueQueueHandler = new DataValueQueueHandler(Id, IsDurable, m_monitoredItemQueueFactory, QueueOverflowHandler);
                             queueLastValue = true;
                         }
 
-                        m_queueHandler.SetQueueSize(m_queueSize, m_discardOldest, m_diagnosticsMasks);
-                        m_queueHandler.SetSamplingInterval(m_samplingInterval);
+                        m_dataValueQueueHandler.SetQueueSize(m_queueSize, m_discardOldest, m_diagnosticsMasks);
+                        m_dataValueQueueHandler.SetSamplingInterval(m_samplingInterval);
 
                         if (queueLastValue && m_lastValue != null)
                         {
-                            m_queueHandler.QueueValue(m_lastValue, m_lastError);
+                            m_dataValueQueueHandler.QueueValue(m_lastValue, m_lastError);
                         }
                     }
                     else // create event queue.
                     {
                         if (m_eventQueueHandler == null)
                         {
-                            m_eventQueueHandler = new EventQueueHandler(Id, IsDurable, m_monitoredItemQueueFactory);
+                            m_eventQueueHandler = new EventQueueHandler(IsDurable, m_monitoredItemQueueFactory);
                         }
                         m_eventQueueHandler.SetQueueSize(m_queueSize, m_discardOldest);
                     }
@@ -1736,8 +1736,8 @@ namespace Opc.Ua.Server
         /// Disposes the durable monitoredItemQueue
         public void Dispose()
         {
-            m_queueHandler?.Dispose();
-            m_eventQueueHandler?.Dispose();
+            Utils.SilentDispose(m_dataValueQueueHandler);
+            Utils.SilentDispose(m_eventQueueHandler);
         }
 
         #endregion
@@ -1772,7 +1772,7 @@ namespace Opc.Ua.Server
         private ServiceResult m_lastError;
         private long m_nextSamplingTime;
         private readonly IMonitoredItemQueueFactory m_monitoredItemQueueFactory;
-        private QueueHandler m_queueHandler;
+        private DataValueQueueHandler m_dataValueQueueHandler;
         private EventQueueHandler m_eventQueueHandler;
         private bool m_readyToPublish;
         private bool m_readyToTrigger;
@@ -1784,349 +1784,5 @@ namespace Opc.Ua.Server
         private bool m_triggered;
         private bool m_resendData;
         #endregion
-
-        #region QueueHandler
-        /// <summary>
-        /// Handles all queue / dequeue actions for a monitoredItem
-        /// </summary>
-        private class QueueHandler : IDisposable
-        {
-            public QueueHandler(uint monitoredItemId, bool createDurable, IMonitoredItemQueueFactory queueFactory, Action discardedValueHandler = null)
-            {
-                m_dataValueQueue = queueFactory.CreateDataValueQueue(createDurable);
-
-                m_discardedValueHandler = discardedValueHandler;
-                m_monitoredItemId = monitoredItemId;
-                m_discardOldest = false;
-                m_nextSampleTime = 0;
-                m_samplingInterval = 0;
-            }
-
-            /// <summary>
-            /// Sets the queue size.
-            /// </summary>
-            /// <param name="queueSize">The new queue size.</param>
-            /// <param name="discardOldest">Whether to discard the oldest values if the queue overflows.</param>
-            /// <param name="diagnosticsMasks">Specifies which diagnostics which should be kept in the queue.</param>
-            public void SetQueueSize(uint queueSize, bool discardOldest, DiagnosticsMasks diagnosticsMasks)
-            {
-                bool queueErrors = (diagnosticsMasks & DiagnosticsMasks.OperationAll) != 0;
-                m_dataValueQueue.SetQueueSize(queueSize, queueErrors);
-
-                // update internals.
-                m_discardOldest = discardOldest;
-            }
-            public void SetSamplingInterval(double samplingInterval)
-            {
-                // substract the previous sampling interval.
-                if (m_samplingInterval < m_nextSampleTime)
-                {
-                    m_nextSampleTime -= m_samplingInterval;
-                }
-
-                // calculate the next sampling interval.
-                m_samplingInterval = (long)samplingInterval;
-
-                if (m_samplingInterval > 0)
-                {
-                    m_nextSampleTime += m_samplingInterval;
-                }
-                else
-                {
-                    m_nextSampleTime = 0;
-                }
-            }
-
-            public int ItemsInQueue => m_dataValueQueue.ItemsInQueue;
-
-            public void QueueValue(DataValue value, ServiceResult error)
-            {
-                long now = HiResClock.TickCount64;
-
-                if (m_dataValueQueue.ItemsInQueue > 0)
-                {
-                    // check if too soon for another sample.
-                    if (now < m_nextSampleTime)
-                    {
-                        m_dataValueQueue.OverwriteLastValue(value, error);
-
-                        m_discardedValueHandler?.Invoke();
-
-                        return;
-                    }
-                }
-
-                // update next sample time.
-                if (m_nextSampleTime > 0)
-                {
-                    long delta = now - m_nextSampleTime;
-
-                    if (m_samplingInterval > 0 && delta >= 0)
-                    {
-                        m_nextSampleTime += ((delta / m_samplingInterval) + 1) * m_samplingInterval;
-                    }
-                }
-                else
-                {
-                    m_nextSampleTime = now + m_samplingInterval;
-                }
-
-                // queue next value.
-                Enqueue(value, error);
-                ServerUtils.EventLog.EnqueueValue(value.WrappedValue);
-            }
-
-            public bool Publish(out DataValue value, out ServiceResult error)
-            {
-                value = null;
-                error = null;
-                if (m_dataValueQueue.Dequeue(out value, out error))
-                {
-                    ServerUtils.EventLog.DequeueValue(value.WrappedValue, value.StatusCode);
-
-                    return true;
-                }
-                return false;
-            }
-
-            private void Enqueue(DataValue value, ServiceResult error)
-            {
-                // check for empty queue.
-                if (m_dataValueQueue.ItemsInQueue == 0)
-                {
-                    ServerUtils.EventLog.EnqueueValue(value.WrappedValue);
-
-                    m_dataValueQueue.QueueValue(value, error);
-
-                    return;
-                }
-
-
-                // check if the latest value has initial dummy data
-                if (m_dataValueQueue.PeekLastValue()?.StatusCode != StatusCodes.BadWaitingForInitialData)
-                {
-                    // check if queue is full.
-                    if (m_dataValueQueue.ItemsInQueue == m_dataValueQueue.QueueSize)
-                    {
-                        m_discardedValueHandler?.Invoke();
-
-                        SetOverflowBit(ref value, ref error);
-
-                        if (!m_discardOldest)
-                        {
-                            ServerUtils.ReportDiscardedValue(null, m_monitoredItemId, value);
-
-                            // overwrite last value
-                            m_dataValueQueue.OverwriteLastValue(value, error);
-
-                            return;
-                        }
-
-                        // remove oldest value.
-                        m_dataValueQueue.Dequeue(out var discardedValue, out var _);
-                        ServerUtils.ReportDiscardedValue(null, m_monitoredItemId, discardedValue);
-                    }
-                    else
-                    {
-                        ServerUtils.EventLog.EnqueueValue(value.WrappedValue);
-                    }
-                }
-                else
-                {
-                    // overwrite the last value
-                    m_dataValueQueue.OverwriteLastValue(value, error);
-                }
-
-                m_dataValueQueue.QueueValue(value, error);
-            }
-
-            /// <summary>
-            /// Sets the overflow bit in the value and error.
-            /// </summary>
-            /// <param name="value">The value to update.</param>
-            /// <param name="error">The error to update.</param>
-            private void SetOverflowBit(ref DataValue value, ref ServiceResult error)
-            {
-                if (value != null)
-                {
-                    StatusCode status = value.StatusCode;
-                    status.Overflow = true;
-                    value.StatusCode = status;
-                }
-
-                if (error != null)
-                {
-                    StatusCode status = error.StatusCode;
-                    status.Overflow = true;
-
-                    // have to copy before updating because the ServiceResult is invariant.
-                    ServiceResult copy = new ServiceResult(
-                        status,
-                        error.SymbolicId,
-                        error.NamespaceUri,
-                        error.LocalizedText,
-                        error.AdditionalInfo,
-                        error.InnerResult);
-
-                    error = copy;
-                }
-            }
-
-            public void Dispose()
-            {
-                m_dataValueQueue.Dispose();
-            }
-
-            private readonly IMonitoredItemQueue<DataValue> m_dataValueQueue;
-            private readonly uint m_monitoredItemId;
-            private bool m_discardOldest;
-            private long m_nextSampleTime;
-            private long m_samplingInterval;
-            private readonly Action m_discardedValueHandler;
-        }
-        #endregion
-
-        #region EventQueueHandler
-        private class EventQueueHandler : IDisposable
-        {
-            public EventQueueHandler(uint monitoredItemId, bool createDurable, IMonitoredItemQueueFactory queueFactory)
-            {
-                m_eventQueue = queueFactory.CreateEventQeue(monitoredItemId, createDurable);
-
-                m_monitoredItemId = monitoredItemId;
-                m_events = new List<EventFieldList>();
-                m_discardOldest = false;
-                m_overflow = false;
-            }
-
-
-            /// <summary>
-            /// Sets the queue size.
-            /// </summary>
-            /// <param name="queueSize">The new queue size.</param>
-            /// <param name="discardOldest">Whether to discard the oldest values if the queue overflows.</param>
-            public void SetQueueSize(uint queueSize, bool discardOldest)
-            {
-                m_discardOldest = discardOldest;
-                m_queueSize = queueSize;
-
-                if (m_events.Count > m_queueSize)
-                {
-                    if (m_discardOldest)
-                    {
-                        m_events.RemoveRange(0, m_events.Count - (int)queueSize);
-                    }
-                    else
-                    {
-                        m_events.RemoveRange((int)queueSize, m_events.Count - (int)queueSize);
-                    }
-                }
-
-                // update internals.
-                m_discardOldest = discardOldest;
-                m_queueSize = queueSize;
-            }
-
-            public int ItemsInQueue => m_events.Count;
-
-            public bool Overflow => m_overflow;
-
-            /// <summary>
-            /// Checks the last 1k queue entries if the event is already in there
-            /// </summary>
-            /// <param name="instance"></param>
-            /// <returns></returns>
-            public bool IsEventContainedInQueue(IFilterTarget instance)
-            {
-                // check for duplicate instances being reported via multiple paths.
-                for (int ii = 0; ii < m_events.Count; ii++)
-                {
-                    if (m_events[ii] is EventFieldList processedEvent)
-                    {
-                        if (ReferenceEquals(instance, processedEvent.Handle))
-                        {
-                            return true;
-                        }
-                    }
-                }
-                return false;
-            }
-
-            /// <summary>
-            /// true if queue is already full and discarding is not allowed
-            /// </summary>
-            /// <returns></returns>
-            public bool SetQueueOverflowIfFull()
-            {
-                if (m_events.Count >= m_queueSize)
-                {
-                    if (!m_discardOldest)
-                    {
-                        m_overflow = true;
-                        return true;
-                    }
-                }
-                return false;
-            }
-            public void Dispose()
-            {
-                m_eventQueue.Dispose();
-            }
-
-            /// <summary>
-            /// Adds an event to the queue.
-            /// </summary>
-            public virtual void QueueEvent(EventFieldList fields)
-            {
-                // make space in the queue.
-                if (m_events.Count >= m_queueSize)
-                {
-                    m_overflow = true;
-
-                    if (m_discardOldest)
-                    {
-                        m_events.RemoveAt(0);
-                    }
-                }
-
-                // queue the event.
-                m_events.Add(fields);
-            }
-
-            public void Publish(OperationContext context, Queue<EventFieldList> notifications)
-            {
-                for (int ii = 0; ii < m_events.Count; ii++)
-                {
-                    EventFieldList fields = m_events[ii];
-
-                    // apply any diagnostic masks.
-                    for (int jj = 0; jj < fields.EventFields.Count; jj++)
-                    {
-                        object value = fields.EventFields[jj].Value;
-
-                        if (value is StatusResult result)
-                        {
-                            result.ApplyDiagnosticMasks(context.DiagnosticsMask, context.StringTable);
-                        }
-                    }
-
-                    notifications.Enqueue(m_events[ii]);
-
-                    m_events.Clear();
-
-                    m_overflow = false;
-                }
-            }
-
-            private bool m_overflow;
-            private uint m_queueSize;
-            private bool m_discardOldest;
-            private List<EventFieldList> m_events;
-            private readonly uint m_monitoredItemId;
-            private readonly IMonitoredItemQueue<EventFieldList> m_eventQueue;
-        }
-        #endregion
     }
-
-
 }
